+++ src/types/Post.ts
@@ -1,4 +1,5 @@
 export interface Post {
+  id: string;
   title: string;
   meetingDate: string;
   location: string;
